+++ ui-src/js/config-fields-service.js
@@ -631,7 +631,7 @@ function ConfigFields($injector) {
                                 label: 'Duplicate age threshold',
                                 required: true,
                                 addonRight: {
-                                    text: 'seconds'
+                                    text: 'hours'
                                 }
                             }
                         },